Move version client to module in mullvad-api
