+++ mullvad-api/src/lib.rs
@@ -3,13 +3,9 @@ use async_trait::async_trait;
 #[cfg(target_os = "android")]
 use futures::channel::mpsc;
 use hyper::body::Incoming;
+use mullvad_types::account::{AccountData, AccountNumber, VoucherSubmission};
 #[cfg(target_os = "android")]
 use mullvad_types::account::{PlayPurchase, PlayPurchasePaymentToken};
-use mullvad_types::{
-    account::{AccountData, AccountNumber, VoucherSubmission},
-    version::AppVersion,
-};
-use mullvad_update::version::{VersionInfo, VersionParameters};
 use proxy::{ApiConnectionMode, ConnectionModeProvider};
 use std::{
     collections::BTreeMap,
@@ -20,11 +16,12 @@ use std::{
     sync::Arc,
 };
 use talpid_types::ErrorExt;
-use vec1::vec1;
 
 pub mod availability;
 use availability::ApiAvailability;
 pub mod rest;
+#[cfg(not(target_os = "ios"))]
+pub mod version;
 
 mod abortable_stream;
 pub mod access_mode;
@@ -696,94 +693,6 @@ impl ProblemReportProxy {
                 .expected_status(&[StatusCode::NO_CONTENT]);
             service.request(request).await?;
             Ok(())
-        }
-    }
-}
-
-#[cfg(not(target_os = "ios"))]
-#[derive(Clone)]
-pub struct AppVersionProxy {
-    handle: rest::MullvadRestHandle,
-}
-
-#[cfg(not(target_os = "ios"))]
-#[derive(serde::Deserialize, Debug)]
-pub struct AppVersionResponse {
-    pub supported: bool,
-    pub latest: AppVersion,
-    pub latest_stable: Option<AppVersion>,
-    pub latest_beta: AppVersion,
-}
-
-#[cfg(not(target_os = "ios"))]
-impl AppVersionProxy {
-    const VERSION_PROVIDER_PUBKEY: &str = include_str!("../../mullvad-update/stagemole-pubkey");
-
-    pub fn new(handle: rest::MullvadRestHandle) -> Self {
-        Self { handle }
-    }
-
-    pub fn version_check(
-        &self,
-        app_version: AppVersion,
-        platform: &str,
-        platform_version: String,
-    ) -> impl Future<Output = Result<AppVersionResponse, rest::Error>> + use<> {
-        let service = self.handle.service.clone();
-
-        let path = format!("{APP_URL_PREFIX}/releases/{platform}/{app_version}");
-        let request = self.handle.factory.get(&path);
-
-        async move {
-            let request = request?
-                .expected_status(&[StatusCode::OK])
-                .header("M-Platform-Version", &platform_version)?;
-            let response = service.request(request).await?;
-            response.deserialize().await
-        }
-    }
-
-    /// Get versions from `/app/releases/<platform>.json`
-    pub fn version_check_2(
-        &self,
-        platform: &str,
-        architecture: mullvad_update::format::Architecture,
-        rollout: f32,
-        lowest_metadata_version: usize,
-    ) -> impl Future<Output = Result<VersionInfo, rest::Error>> + use<> {
-        // Maximum size of version response
-        const SIZE_LIMIT: usize = 1024 * 1024;
-
-        let service = self.handle.service.clone();
-        let path = format!("app/releases/{platform}.json");
-        let request = self.handle.factory.get(&path);
-
-        let verifying_key =
-            mullvad_update::format::key::VerifyingKey::from_hex(Self::VERSION_PROVIDER_PUBKEY)
-                .expect("valid key");
-        let verifying_keys = vec1![verifying_key];
-
-        async move {
-            let request = request?.expected_status(&[StatusCode::OK]);
-            let response = service.request(request).await?;
-            let bytes = response.body_with_max_size(SIZE_LIMIT).await?;
-
-            let response = mullvad_update::format::SignedResponse::deserialize_and_verify(
-                &verifying_keys,
-                &bytes,
-                lowest_metadata_version,
-            )
-            .map_err(|err| rest::Error::FetchVersions(Arc::new(err)))?;
-
-            let params = VersionParameters {
-                architecture,
-                rollout,
-                lowest_metadata_version,
-            };
-
-            VersionInfo::try_from_response(&params, response.signed)
-                .map_err(Arc::new)
-                .map_err(rest::Error::FetchVersions)
         }
     }
 }
